requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.2
django-crispy-forms==2.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
